make navigation ssr: false
