+++ apps/chat/src/components/Profile/User/UserMobile.tsx
@@ -4,13 +4,13 @@ import { Translation } from '@/src/types/translation';
 
 import { UserAvatar } from './UserAvatar';
 
-export const ProfileButton = () => {
+export const UserMobile = () => {
   const { t } = useTranslation(Translation.Header);
 
   return (
     <div
       className="flex items-center justify-center text-secondary md:text-primary"
-      data-qa="account-settings"
+      data-qa="account-settings-mobile"
       aria-label={t('Account settings')}
     >
       <UserAvatar />
